Refactor special character encoding
When encountering special characters in an Excel file, readxl
represents them as UTF-8. GDXXRW however uses a latin code page,
probably windows-1252. So the logical thing to do is to try to encode
special characters with windows-1252 before generating the GDX.
However, in spite of this encoding being defined on Linux (see
stri_enc_list()), doing such conversion can result in a "bytes"
encoding. Hence, the more standard and mostly equivalent (other than
in the 0x80-0x9F range) ISO-8859-1 encoding was tried instead, but it
can result in a "bytes" encoding on Windows. Therefore, first one and
then the other encoding is tried. If either one succeeds on a given
platform, the results are likely the same unless one of the weird
characters in the 0x80-0x9F range is present in the input. For a list
of windows-1252 vs ISO-8859-1 differences see
https://en.wikipedia.org/wiki/Windows-1252.